cleaned data

+++ db/migrate/20240722171401_create_air_quality_data.rb
@@ -1,10 +1,10 @@
 class CreateAirQualityData < ActiveRecord::Migration[7.1]
   def change
     create_table :air_quality_data do |t|
-      t.integer :aqi
-      t.integer :temp
-      t.integer :humidity
-      t.integer :wspd
+      t.float :aqi
+      t.float :temp
+      t.float :humidity
+      t.float :wspd
 
       t.timestamps
     end
