+++ db/migrate/20240722181009_add_details_to_air_quality_data.rb
@@ -1,7 +1,7 @@
 class AddDetailsToAirQualityData < ActiveRecord::Migration[7.1]
   def change
-    add_column :air_quality_data, :temp, :integer
-    add_column :air_quality_data, :humidity, :integer
-    add_column :air_quality_data, :wspd, :integer
+    add_column :air_quality_data, :temp, :float
+    add_column :air_quality_data, :humidity, :float
+    add_column :air_quality_data, :wspd, :float
   end
 end
